Case 1928: Import/Export Blobs
git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@2495 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2

Removed customer's generated code and schema

+++ Kistl.App.Projekte.Common/CustomCommonActions_DocumentManagement.cs
@@ -16,7 +16,7 @@ namespace at.dasz.DocumentManagement
         public static void OnHandleBlobChange_DynamicFile(at.dasz.DocumentManagement.DynamicFile obj, MethodReturnEventArgs<Kistl.App.Base.Blob> e, Kistl.App.Base.Blob oldBlob, Kistl.App.Base.Blob newBlob)
         {
             // Delete old blob
-            if (oldBlob != null)
+            if (oldBlob != null && newBlob != oldBlob)
             {
                 obj.Context.Delete(oldBlob);
             }
